refactor index.js -> index.ts + add TODO/@debt comments
